Fix formatting of second-level bullet points
Nested list items are indented 4 spaces in Kramdown:

https://kramdown.gettalong.org/syntax.html#ordered-and-unordered-lists

However when I replace `⋅⋅` with `    * ` in these lines, the numbers in the
nested lists all render as "1" - somehow these lines seem to be being
interpreted as ordered and unordered lists simultaneously, but with broken
numbering. Replacing the `.` after the number with `:` seems to preserve the
correct numbering after the bullets.

Removing the bullets from the nested lists allows the numbers to render
correctly, but the indentation levels then look wrong because the nested lists
are then aligned with the text after the bullets in the outer lists. This
makes it hard to see the nesting clearly and looks odd.

An alternative approach would be to remove the bullets entirely from all
lists, and only use letters and numbers. I think that looks less clear when
rendered - the bullets are useful for visually identifying lists in these
pages.

For consistency, we could use `:` after the letters in the outer lists too if
that's clearer and easier to use for people editing this content.