Update leak-handling-process.md
Hi Adrian, I've gone through this content and it seems complete and simplified but the ordering of the steps have changed. Briefly, this version and my process/RACi doc align up to step 2.1.1, but deviate going forwards. In my doc the internal investigation(before escalation) and the creation of the ServiceNow ticket are in parallel, whereas in this version it is only after escalation (formal investigation) is confirmed the ServiceNow ticket is created.  I take it, this is intentional?  Have carried out snagging on doc for approval.