Queries on wording used, removal of links to wikipedia
Under the heading of Testing on line 20 John has queried the must on line 22.5 is correct.

On line 52 with the heading of End-user device based John has advised that Perhaps remove wikipedia links as Wikipedia is an arbitrary site editable by anyone. I have included some different ones to consider. I'm not too sure on the bitlocker one as I thought these were gotten rid of now. 

On line 68 the link to NATO Restricted Level Certified was a dead link. I tried looking for a new one but the websites that came back seemed a bit dodgy. 

John suggested merging the sections of portable end-user devices and end-user device based into one