Update enterprise-it-mobile-devices.md
John has commented stating overall this page looks ok but in draft mode, eg spelling, punctuation, style “to be added” etc. Suggest refer back to document author for QA / query current relevance.

I have tried my best to re-jigg. To be honest it's a bit above my head. I noticed that the log collection principle 3 was missing, not sure if maybe it not applicable in this or not.