Review by creator
John comments are the same as the enterprise of IT mobile phones one. Overall this page looks ok but in draft mode, eg spelling, punctuation, style “to be added” etc. Suggest refer back to document author for QA / query current relevance

Have done some tweaking but think it will need more with someone with the knowlegde.